+++ domain/object_position.go
@@ -1,8 +1,0 @@
-package domain
-
-type Position struct {
-	X      int
-	Y      int
-	Width  int
-	Height int
-}