+++ engine/update_objects.go
@@ -2,80 +2,114 @@ package engine
 
 import (
 	"mini-game-go/domain"
+	"strconv"
 	"time"
 
 	"github.com/hajimehoshi/ebiten/v2"
 )
 
-func updateRoad(g *Game) {
-	if g.gameOver.Flag {
+func updateRoad(game *Game) {
+	if game.gameOver.Flag {
 		return
 	}
-	if g.car.Fuel.Percent == 0 {
-		g.gameOver.TextPosition.Y = 300
-		g.gameOver.BoxPosition.Y = 300
-
-		g.gameOver.Flag = true
+	if time.Since(game.roadMove) >= 50*time.Millisecond {
+		game.roadMove = time.Now()
+		moveLinesRoad(game)
+		moveObstaclesRoad(game)
 	}
-	if time.Since(g.roadMove) >= 50*time.Millisecond {
-		g.roadMove = time.Now()
-		// Move road
-		for i := 0; i < len(g.road); i++ {
-			g.road[i].Y += g.car.Speed * 2
-			if g.road[i].Y > int(domain.GameHeight) {
-				g.road[i].Y = -g.road[i].Height - 50
-			}
+}
+
+func moveLinesRoad(game *Game) {
+	for i := 0; i < len(game.road); i++ {
+		game.road[i].Position.Y += game.car.Speed * 2
+		if game.road[i].Position.Y > domain.GameHeightInt() {
+			game.road[i].Position.Y = -game.road[i].Size.Height - 50
 		}
-		for i := 0; i < len(g.obstacles); i++ {
-			g.obstacles[i].Position.Y += g.car.Speed * 2
-			if g.obstacles[i].Position.Y > int(domain.GameHeight) {
-				g.obstacles = append(g.obstacles[:i], g.obstacles[i+1:]...)
-				obstaclesGame, _ := loadObstacles(1, g)
-				g.obstacles = append(g.obstacles, obstaclesGame[0])
-			}
+	}
+}
+func moveObstaclesRoad(game *Game) {
+	for i := 0; i < len(game.obstacles); i++ {
+		game.obstacles[i].Object.Position.Y += game.car.Speed * 2
+		if game.obstacles[i].Object.Position.Y > domain.GameHeightInt() {
+			game.obstacles = append(game.obstacles[:i], game.obstacles[i+1:]...)
+			obstaclesGame, _ := loadObstacles(1, game)
+			game.obstacles = append(game.obstacles, obstaclesGame[0])
 		}
 	}
 }
 
-func updateFuel(g *Game) {
-	if g.gameOver.Flag {
+func updateFuel(game *Game) {
+	if game.gameOver.Flag {
 		return
 	}
-	colorFuel := getColorFuel(g.car.Fuel.Percent)
-	if colorFuel == "" {
-		g.gameOver.Flag = false
-	}
-	g.car.Fuel.Color = colorFuel
+	colorFuel := getColorFuel(game.car.Fuel.Percent)
+	game.car.Fuel.Color = colorFuel
 
-	if time.Since(g.car.Fuel.Time) >= 1000*time.Millisecond {
-		g.car.Fuel.Time = time.Now()
-		fuelConsume := int(g.car.Speed / 5)
-		if fuelConsume <= 0 {
-			fuelConsume = 1
+	if time.Since(game.car.Fuel.Time) >= 1000*time.Millisecond {
+		game.car.Fuel.Time = time.Now()
+		taxConsume := int(game.car.Speed / 5)
+		if taxConsume <= 0 {
+			taxConsume = 1
 		}
-		g.car.Fuel.Percent -= fuelConsume
-		if g.car.Fuel.Percent < 0 {
-			g.car.Fuel.Percent = 0
-			g.gameOver.Flag = true
+		game.car.Fuel.Percent -= taxConsume
+		if game.car.Fuel.Percent <= 0 {
+			game.car.Fuel.Percent = 0
+			drawGameOver(game)
 		}
 	}
 }
 
-func updateCar(g *Game) {
-	if g.gameOver.Flag {
+func verifyConflict(game *Game, mainObject, conflictObject domain.Object, margin int) bool {
+	// mainObject = car
+	// conflictObject = obstacle
+	confiltoSupDir := mainObject.Position.X+mainObject.Size.Width+margin <= conflictObject.Position.X+conflictObject.Size.Width
+	confiltoSupEsc := mainObject.Position.X+margin > conflictObject.Position.X
+	if confiltoSupDir && confiltoSupEsc {
+		if mainObject.Position.Y-margin <= conflictObject.Position.Y+conflictObject.Size.Height {
+			game.gameOver.Text = strconv.FormatBool(confiltoSupDir) + "|" + strconv.FormatBool(confiltoSupEsc)
+			return true
+		}
+	}
+	return false
+}
+
+func updateCar(game *Game) {
+	if game.gameOver.Flag {
 		return
+	}
+	for i := 0; i < len(game.obstacles); i++ {
+		if verifyConflict(game, game.car.Object, game.obstacles[i].Object, 0) {
+			drawGameOver(game)
+		}
 	}
 	//  update position main car
-	if (ebiten.IsKeyPressed(ebiten.KeyArrowLeft) || ebiten.IsKeyPressed(ebiten.KeyA)) && g.car.Position.X >= 5 {
-		g.car.Position.X -= int(g.car.Speed)
+	if (ebiten.IsKeyPressed(ebiten.KeyArrowLeft) || ebiten.IsKeyPressed(ebiten.KeyA)) &&
+		game.car.Object.Position.X >= 20 {
+		game.car.Object.Position.X -= int(game.car.Speed)
 	}
-	if (ebiten.IsKeyPressed(ebiten.KeyArrowRight) || ebiten.IsKeyPressed(ebiten.KeyD)) && g.car.Position.X <= 690 {
-		g.car.Position.X += int(g.car.Speed)
+	if (ebiten.IsKeyPressed(ebiten.KeyArrowRight) || ebiten.IsKeyPressed(ebiten.KeyD)) &&
+		game.car.Object.Position.X <= domain.GameWidthInt()-game.car.Object.Size.Width-20 {
+		game.car.Object.Position.X += int(game.car.Speed)
 	}
-	if (ebiten.IsKeyPressed(ebiten.KeyArrowUp) || ebiten.IsKeyPressed(ebiten.KeyW)) && g.car.Position.Y >= 55 {
-		g.car.Position.Y -= int(g.car.Speed)
+	if (ebiten.IsKeyPressed(ebiten.KeyArrowUp) || ebiten.IsKeyPressed(ebiten.KeyW)) &&
+		game.car.Object.Position.Y >= 60 {
+		game.car.Object.Position.Y -= int(game.car.Speed)
 	}
-	if (ebiten.IsKeyPressed(ebiten.KeyArrowDown) || ebiten.IsKeyPressed(ebiten.KeyS)) && g.car.Position.Y <= 757 {
-		g.car.Position.Y += int(g.car.Speed)
+	if (ebiten.IsKeyPressed(ebiten.KeyArrowDown) || ebiten.IsKeyPressed(ebiten.KeyS)) &&
+		game.car.Object.Position.Y <= domain.GameWidthInt()-20 {
+		game.car.Object.Position.Y += int(game.car.Speed)
 	}
+	if ebiten.IsKeyPressed(ebiten.KeySpace) {
+		if game.car.Speed == 0 {
+			game.car.Speed = 5
+		} else {
+			game.car.Speed = 0
+		}
+	}
+}
+
+func drawGameOver(game *Game) {
+	game.gameOver.TextPosition.Y = 300
+	game.gameOver.BoxObject.Position.Y = 300
+	game.gameOver.Flag = true
 }